+++ package.json
@@ -63,7 +63,8 @@
     "prettier": "1.17.1",
     "pretty-quick": "^1.11.0",
     "react-test-renderer": "^16.8.6",
-    "redux-devtools-extension": "^2.13.8"
+    "redux-devtools-extension": "^2.13.8",
+    "redux-mock-store": "^1.5.3"
   },
   "husky": {
     "hooks": {
